sq - move searches in searches folder
